Feat: View Model 내부 코드 수정
