+++ feature-userinfo/src/main/java/com/example/feature_userinfo/viewModel/PostViewModel.kt
@@ -1,7 +1,10 @@
 package com.example.feature_userinfo.viewModel
 
+import com.hmoa.core_model.response.CommunityByCategoryResponseDto
 import com.hmoa.core_repository.Member.MemberRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
 import javax.inject.Inject
 
 @HiltViewModel
@@ -9,6 +12,9 @@ class PostViewModel @Inject constructor(
     private val repository : MemberRepository
 ){
 
-
+    //작성한 게시글
+    private var _postList : Flow<List<CommunityByCategoryResponseDto>> = flowOf(emptyList())
+    val postList : Flow<List<CommunityByCategoryResponseDto>>
+        get() = _postList
 
 }
